requirements.txt
tk-tools
win10toast
mysql-connector-python